fix (partially) DirtyQ
DirtyQ was evaluating its argument normally.
I added the attribute HoldAll, and evaluate the
argument once before calling checkdirtysyms.
This seems to be more useful. It does what it claims
to do. But, subsequent evaluations of the expression
do not change its age. This is not the behavior we want.
	modified:   src/mxpr.jl
	modified:   src/predicates.jl
	modified:   src/protected_symbols.jl

* fix typo in notebook