Use x!num for Module-local variables
That is x!1, x!2, etc. This behaves like Mma. But, the Module implementation
is still broken. The variables are supposed to be sort-of lexically scoped.
Except they are allowed to leak out. There is a Module-local counter that
increases every time a set of variables is created. We made module more efficient
and broke this last behavior in the process. Instead the local variable is
created only when Module is first evaluated. It is cleared on subsequent calls.
Fixes #164

	modified:   src/evaluation.jl
	modified:   src/kernelstate.jl
	modified:   src/module.jl